@@ -1,49 +1,60 @@
-defaultTasks 'clean','test','aggregate'
+defaultTasks 'clean', 'test', 'aggregate'
 
 repositories {
+    mavenCentral()
     mavenLocal()
-    jcenter()
 }
 
 buildscript {
     repositories {
-        mavenLocal()
-        jcenter()
+        maven {
+            url "https://plugins.gradle.org/m2/"
+        }
     }
     dependencies {
-        classpath("net.serenity-bdd:serenity-gradle-plugin:2.2.28")
+        classpath "net.serenity-bdd:serenity-gradle-plugin:4.0.15"
     }
 }
+
 
 apply plugin: 'java'
 apply plugin: 'eclipse'
 apply plugin: 'idea'
-apply plugin: 'net.serenity-bdd.aggregator'
+apply plugin: "net.serenity-bdd.serenity-gradle-plugin"
 
-sourceCompatibility = 1.8
-targetCompatibility = 1.8
+sourceCompatibility = 16
+targetCompatibility = 16
 
 ext {
-    slf4jVersion = '1.7.7'
-    serenityCoreVersion = '2.2.28'
-    junitVersion = '4.13'
-    assertJVersion = '3.8.0'
-    logbackVersion = '1.2.3'
+    serenity_version = '4.0.15'
+    junit_platform_launcher_version="1.10.0"
+    cucumber_junit_platform_engine_version="7.14.0"
+    junit_platform_suite_version="1.10.0"
+    junit_jupiter_engine_version="5.9.3"
+    junit_vintage_engine_version="5.9.3"
+    logback_classic_version="1.2.10"
+    assertj_core_version="3.23.1"
 }
 
 dependencies {
-    compile "ch.qos.logback:logback-classic:${logbackVersion}"
-
-    testCompile "net.serenity-bdd:serenity-core:${serenityCoreVersion}",
-                "net.serenity-bdd:serenity-junit:${serenityCoreVersion}",
-                "net.serenity-bdd:serenity-screenplay:${serenityCoreVersion}",
-                "junit:junit:${junitVersion}",
-                "org.assertj:assertj-core:${assertJVersion}"
+    dependencies {
+        testImplementation "net.serenity-bdd:serenity-core:${serenity_version}"
+        testImplementation "net.serenity-bdd:serenity-cucumber:${serenity_version}"
+        testImplementation "net.serenity-bdd:serenity-screenplay:${serenity_version}"
+        testImplementation "net.serenity-bdd:serenity-screenplay-webdriver:${serenity_version}"
+        testImplementation "net.serenity-bdd:serenity-ensure:${serenity_version}"
+        testImplementation "org.junit.platform:junit-platform-launcher:${junit_platform_launcher_version}"
+        testImplementation "io.cucumber:cucumber-junit-platform-engine:${cucumber_junit_platform_engine_version}"
+        testImplementation "org.junit.platform:junit-platform-suite:${junit_platform_suite_version}"
+        testImplementation "org.junit.jupiter:junit-jupiter-engine:${junit_jupiter_engine_version}"
+        testImplementation "org.junit.vintage:junit-vintage-engine:${junit_vintage_engine_version}"
+        implementation "ch.qos.logback:logback-classic:${logback_classic_version}"
+        testImplementation "org.assertj:assertj-core:${assertj_core_version}"
+    }
 }
 
 test {
-    testLogging.showStandardStreams = true
-    systemProperties System.getProperties()
+    useJUnitPlatform()
 }
 
 gradle.startParameter.continueOnFailure = true
