@@ -1,6 +1,5 @@
 package gateway.test;
 
-import gateway.helper.SeleniumBase;
 import gateway.pages.VanillaGatewayPage;
 import org.junit.Test;
 
@@ -12,7 +11,6 @@ public class VanillaGatewayTests extends SeleniumBase {
 
         VanillaGatewayPage gateway = new VanillaGatewayPage(getDriver());
         gateway.open().clickCookiesAcceptButton().clickSendButton();
-        //gateway.clickSomeButton();
-        //gateway.clickSendButton();
+        //etc
     }
 }
